# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface Credentials {
  email: string;
  password: string;
}

/**
 * Saved app state snapshot (null if none saved yet)
 */
export type AuthUserState = { [key: string]: unknown } | null;

export interface AuthUser {
  email: string;
  /** Saved app state snapshot (null if none saved yet) */
  state?: AuthUserState;
}

/**
 * Full app state snapshot to persist
 */
export type StateSaveRequestState = { [key: string]: unknown };

export interface StateSaveRequest {
  /** Full app state snapshot to persist */
  state: StateSaveRequestState;
}

export interface HealthStatus {
  status: string;
}

/**
 * Internal recovery stage (never shown to user)
 */
export type ChallengeRequestStage = typeof ChallengeRequestStage[keyof typeof ChallengeRequestStage];


export const ChallengeRequestStage = {
  secluded: 'secluded',
  highly_isolated: 'highly_isolated',
  at_risk: 'at_risk',
  not_at_risk: 'not_at_risk',
} as const;

/**
 * Challenge area for today
 */
export type ChallengeRequestArea = typeof ChallengeRequestArea[keyof typeof ChallengeRequestArea];


export const ChallengeRequestArea = {
  rhythm: 'rhythm',
  selfcare: 'selfcare',
  relationship: 'relationship',
  social: 'social',
} as const;

export interface ChallengeRequest {
  /** Internal recovery stage (never shown to user) */
  stage: ChallengeRequestStage;
  /** Challenge area for today */
  area: ChallengeRequestArea;
  /**
     * Lowest difficulty level for today's set
     * @minimum 1
     * @maximum 5
     */
  bandLow: number;
  /**
     * Highest difficulty level for today's set
     * @minimum 1
     * @maximum 5
     */
  bandHigh: number;
  /** Forbidden action types (e.g. outing, face-to-face, phone-call) */
  forbidden: string[];
  /** Onboarding sleep pattern answer */
  sleep: string;
  /** Onboarding outing burden answer */
  outing: string;
  /** Onboarding contact comfort answer */
  contact: string;
  /** Today's condition answer */
  condition: string;
  /** Today's interest/preference answer */
  interest: string;
}

export type ChallengeArea = typeof ChallengeArea[keyof typeof ChallengeArea];


export const ChallengeArea = {
  rhythm: 'rhythm',
  selfcare: 'selfcare',
  relationship: 'relationship',
  social: 'social',
} as const;

export interface Challenge {
  area: ChallengeArea;
  /**
     * @minimum 1
     * @maximum 5
     */
  level: number;
  /** Challenge phrase shown on the card */
  title: string;
  /** Estimated minutes to complete */
  minutes: number;
  /** One-line reflection question shown after completion */
  reflectQ: string;
}

/**
 * Whether challenges came from the LLM or the fallback bank
 */
export type ChallengeSetSource = typeof ChallengeSetSource[keyof typeof ChallengeSetSource];


export const ChallengeSetSource = {
  llm: 'llm',
  fallback: 'fallback',
} as const;

export interface ChallengeSet {
  challenges: Challenge[];
  /** Whether challenges came from the LLM or the fallback bank */
  source: ChallengeSetSource;
}