# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: challenges
    description: Challenge generation
  - name: auth
    description: Authentication and user state persistence
paths:
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create an account and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Account created and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid input
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Wrong email or password
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the current session
      responses:
        "204":
          description: Logged out
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current logged-in user and saved app state
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not logged in
  /state:
    put:
      operationId: saveState
      tags: [auth]
      summary: Save the user's app state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StateSaveRequest"
      responses:
        "204":
          description: Saved
        "401":
          description: Not logged in
  /challenges/generate:
    post:
      operationId: generateChallenges
      tags: [challenges]
      summary: Generate today's 3 challenges via LLM
      description: >
        Takes the rule-based classifier output (stage, area, difficulty band, forbidden actions)
        plus user context, and returns 3 personalized challenges generated by an LLM.
        Falls back to the hardcoded example bank on failure or timeout.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeRequest"
      responses:
        "200":
          description: Three challenges (low, medium, high difficulty within the band)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChallengeSet"
        "400":
          description: Invalid request
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  schemas:
    Credentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [email]
      properties:
        email:
          type: string
        state:
          type: object
          description: Saved app state snapshot (null if none saved yet)
          additionalProperties: true
          nullable: true
    StateSaveRequest:
      type: object
      required: [state]
      properties:
        state:
          type: object
          description: Full app state snapshot to persist
          additionalProperties: true
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ChallengeRequest:
      type: object
      required: [stage, area, bandLow, bandHigh, forbidden, sleep, outing, contact, condition, interest]
      properties:
        stage:
          type: string
          description: Internal recovery stage (never shown to user)
          enum: [secluded, highly_isolated, at_risk, not_at_risk]
        area:
          type: string
          description: Challenge area for today
          enum: [rhythm, selfcare, relationship, social]
        bandLow:
          type: integer
          minimum: 1
          maximum: 5
          description: Lowest difficulty level for today's set
        bandHigh:
          type: integer
          minimum: 1
          maximum: 5
          description: Highest difficulty level for today's set
        forbidden:
          type: array
          items:
            type: string
          description: Forbidden action types (e.g. outing, face-to-face, phone-call)
        sleep:
          type: string
          description: Onboarding sleep pattern answer
        outing:
          type: string
          description: Onboarding outing burden answer
        contact:
          type: string
          description: Onboarding contact comfort answer
        condition:
          type: string
          description: Today's condition answer
        interest:
          type: string
          description: Today's interest/preference answer
    Challenge:
      type: object
      required: [area, level, title, minutes, reflectQ]
      properties:
        area:
          type: string
          enum: [rhythm, selfcare, relationship, social]
        level:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
          description: Challenge phrase shown on the card
        minutes:
          type: integer
          description: Estimated minutes to complete
        reflectQ:
          type: string
          description: One-line reflection question shown after completion
    ChallengeSet:
      type: object
      required: [challenges, source]
      properties:
        challenges:
          type: array
          items:
            $ref: "#/components/schemas/Challenge"
        source:
          type: string
          enum: [llm, fallback]
          description: Whether challenges came from the LLM or the fallback bank