# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserState } from './authUserState';

export interface AuthUser {
  email: string;
  /** Saved app state snapshot (null if none saved yet) */
  state?: AuthUserState;
}