# lib/api-zod/src/generated/types/stateSaveRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StateSaveRequestState } from './stateSaveRequestState';

export interface StateSaveRequest {
  /** Full app state snapshot to persist */
  state: StateSaveRequestState;
}